less jitter due to translation when rendering objects
Former-commit-id: 83d6567eda39a92d76b52d805c50a16d3e44e783

+++ Engine/ComponentSystem/Components/StaticPhysicsRenderer.cs
@@ -33,8 +33,8 @@ namespace Engine.ComponentSystem.Components
             // Draw the texture based on our physics component.
             p.SpriteBatch.Begin();
             p.SpriteBatch.Draw(texture,
-                new Rectangle((int)((float)sphysics.Position.X + p.Translation.X),
-                              (int)((float)sphysics.Position.Y + p.Translation.Y),
+                new Rectangle((int)sphysics.Position.X + (int)p.Translation.X,
+                              (int)sphysics.Position.Y + (int)p.Translation.Y,
                               texture.Width, texture.Height),
                 null, Color.White,
                 (float)sphysics.Rotation,
